Add pre-requisite of using a unix-based system
Python and Flask are much easier to use on a Unix based system. This is due to
better developer support, both in terms of online forums as well as tooling
(such as better package managers and other command-line tooling). Hence, for
Windows users, it is important to have a Linux subsystem that can be used in
a similar way to a unix command-line.

In addition, by enforcing support only for Unix-based systems, this simplifies
the instructions required to be written.